compclassf.classify {ddalpha}	R Documentation

Classify using Functional Componentwise Classifier

Description

Classifies data using the functional componentwise classifier.

Usage

compclassf.classify(compclassf, objectsf, subset, ...)

## S3 method for class 'compclassf'
predict(object, objectsf, subset, ...)

Arguments

compclassf, object
Functional componentwise classifier (obtained by compclassf.train).

objectsf
list containing lists (functions) of two vectors of equal length, named "args" and "vals": arguments sorted in ascending order and corresponding them values respectively

subset
an optional vector specifying a subset of observations to be classified.

...
additional parameters, passed to the classifier, selected with parameter classifier.type in compclassf.train.

Value

List containing class labels.

References

Delaigle, A., Hall, P., and Bathia, N. (2012). Componentwise classification and clustering of functional data. Biometrika 99 299–313.

See Also

compclassf.train to train the functional componentwise classifier.

Examples

## Not run: 
## load the Growth dataset
dataf = dataf.growth()

learn = c(head(dataf$dataf, 49), tail(dataf$dataf, 34))
labels =c(head(dataf$labels, 49), tail(dataf$labels, 34)) 
test = tail(head(dataf$dataf, 59), 10)    # elements 50:59. 5 girls, 5 boys

c = compclassf.train (learn, labels, classifier.type = "ddalpha")

classified = compclassf.classify(c, test)

print(unlist(classified))

## End(Not run)

[Package ddalpha version 1.3.1.1 Index]