dataf.population {ddalpha}	R Documentation

World Historical Population-by-Country Dataset

Description

Historical world population data by countries.

Usage

dataf.population()

Format

The functional data as a data structure.

dataf

The functional data as a list of objects. Each object is characterized by two coordinates

args

year - a numeric vector of years 1950-2015 (66 years)

vals

population - a numeric vector of the estimated total population in thousands in 233 countries and regions

labels

The geographic region of the country: Africa, Asia, Europe, Latin America, North America, Oceania

identifier

The name of country or region

Details

World population data by a country, area or region as of 1 July of the year indicated. Figures are presented in thousands.

Source

United Nations, Department of Economic and Social Affairs, Population Division, https://esa.un.org/unpd/wpp/Download/Standard/Population/, file Total population - Both sexes

See Also

dataf.population2010

dataf.* for other functional data sets

plot.functional for building plots of functional data

Examples

## load the Population dataset
dataf = dataf.population()

## view the classes
unique(dataf$labels)

## access the 5th point of the 2nd object
dataf$dataf[[2]]$args[5]
dataf$dataf[[2]]$vals[5]

## plot the data
## Not run:
labels = unlist(dataf$labels)
plot(dataf,
  main = "World population data",
  xlab="Year", ylab="Population (in thousands)"
  )
## End(Not run)

## compute the integrated and infimal depths of the data curves
## with respect to the same set of curves
depthf.fd1(dataf$dataf, dataf$dataf)

[Package ddalpha version 1.3.1.1 Index]