dataf.tecator {ddalpha}	R Documentation

Functional Data Set Spectrometric Data (Tecator)

Description

This dataset is a part of the original one which can be found at http://stat.cmu.edu. For each peace of finely chopped meat we observe one spectrometric curve which corresponds to the absorbance measured at 100 wavelengths. The peaces are split according to Ferraty and Vieu (2006) into two classes: with small (<20) and large fat content obtained by an analytical chemical processing.

Usage

dataf.tecator()

Format

The functional data as a data structure.

dataf

The functional data as a list of objects. Each object is characterized by two coordinates.

args

wavelength - a numeric vector of discretization points from 850 to 1050mm

vals

absorbance - a numeric vector of absorbance values

labels

The classes of the objects: "small" (<20) and "large" fat content

Author(s)

Febrero-Bande, M and Oviedo de la Fuente, Manuel

Source

http://stat.cmu.edu

References

Ferraty, F. and Vieu, P. (2006). Nonparametric functional data analysis: theory and practice. Springer.

See Also

dataf.* for other functional data sets

plot.functional for building plots of functional data

Examples

## load the dataset
dataf = dataf.tecator()

## view the classes
unique(dataf$labels)

## access the 5th point of the 2nd object
dataf$dataf[[2]]$args[5]
dataf$dataf[[2]]$vals[5]

## plot the data
## Not run: 
labels = unlist(dataf$labels)
plot(dataf, 
  xlab="Wavelengths", ylab="Absorbances", 
  main=paste("Tecator: < 20 red (", sum(labels == "small"), "),", 
            " >= 20 blue (", sum(labels == "large"), ")", sep=""),
    colors = c("blue", "red"))

## End(Not run)

[Package ddalpha version 1.3.1.1 Index]