ddalpha.test {ddalpha}	R Documentation

Test DD-Classifier

Description

Trains DD-classifier on the learning sequence of the data and tests it on the testing sequence.

Usage

ddalpha.test(learn, test, ...)

Arguments

learn
the learning sequence of the data. Matrix containing training sample where each of n rows is one object of the training sample where first d entries are inputs and the last entry is output (class label).

test
the testing sequence. Has the same format as learn

...
additional parameters passed to ddalpha.train

Value

error
the part of incorrectly classified data

correct
the number of correctly classified objects

incorrect
the number of incorrectly classified objects

total
the number of classified objects

ignored
the number of ignored objects (outside the convex hull of the learning data)

n
the number of objects in the testing sequence

time
training time

See Also

ddalpha.train to train the DD-classifier, ddalpha.classify for classification using DD-classifier, ddalpha.getErrorRateCV and ddalpha.getErrorRatePart to get error rate of the DD-classifier on particular data.

Examples

# Generate a bivariate normal location-shift classification task
# containing 200 training objects and 200 to test with
class1 <- mvrnorm(200, c(0,0), 
                  matrix(c(1,1,1,4), nrow = 2, ncol = 2, byrow = TRUE))
class2 <- mvrnorm(200, c(2,2), 
                  matrix(c(1,1,1,4), nrow = 2, ncol = 2, byrow = TRUE))
trainIndices <- c(1:100)
testIndices <- c(101:200)
propertyVars <- c(1:2)
classVar <- 3
trainData <- rbind(cbind(class1[trainIndices,], rep(1, 100)), 
                   cbind(class2[trainIndices,], rep(2, 100)))
testData <- rbind(cbind(class1[testIndices,], rep(1, 100)), 
                  cbind(class2[testIndices,], rep(2, 100)))
data <- list(train = trainData, test = testData)

# Train 1st DDalpha-classifier (default settings) 
# and get the classification error rate
stat <- ddalpha.test(data$train, data$test)
cat("1. Classification error rate (defaults): ", 
    stat$error, ".\n", sep = "")

# Train 2nd DDalpha-classifier (zonoid depth, maximum Mahalanobis 
# depth classifier with defaults as outsider treatment) 
# and get the classification error rate
stat2 <- ddalpha.test(data$train, data$test, depth = "zonoid", 
                          outsider.methods = "depth.Mahalanobis")
cat("2. Classification error rate (depth.Mahalanobis): ", 
    stat2$error, ".\n", sep = "")

[Package ddalpha version 1.3.1.1 Index]