ddalphaf.classify {ddalpha}	R Documentation

Classify using Functional DD-Classifier

Description

Classifies data using the functional DD-classifier.

Usage

ddalphaf.classify(ddalphaf, objectsf, subset, ...)

## S3 method for class 'ddalphaf'
predict(object, objectsf, subset, ...)

Arguments

ddalphaf, object
Functional DD-classifier (obtained by ddalphaf.train).

objectsf
list containing lists (functions) of two vectors of equal length, named "args" and "vals": arguments sorted in ascending order and corresponding them values respectively

subset
an optional vector specifying a subset of observations to be classified.

...
additional parameters, passed to the classifier, selected with parameter classifier.type in ddalphaf.train.

Value

List containing class labels.

References

Mosler, K. and Mozharovskyi, P. (2015). Fast DD-classification of functional data, Statistical Papers, in press.

Mozharovskyi, P. (2015). Contributions to Depth-based Classification and Computation of the Tukey Depth. Verlag Dr. Kovac (Hamburg).

See Also

ddalphaf.train to train the functional DDα-classifier.

Examples

## Not run: 
## load the Growth dataset
dataf = dataf.growth()

learn = c(head(dataf$dataf, 49), tail(dataf$dataf, 34))
labels= c(head(dataf$labels, 49), tail(dataf$labels, 34)) 
test  = tail(head(dataf$dataf, 59), 10)    # elements 50:59. 5 girls, 5 boys

c = ddalphaf.train (learn, labels, classifier.type = "ddalpha")

classified = ddalphaf.classify(c, test)

print(unlist(classified))

## End(Not run)

[Package ddalpha version 1.3.1.1 Index]