ddalphaf.getErrorRatePart {ddalpha}	R Documentation

Test Functional DD-Classifier

Description

Performs a benchmark procedure by partitioning the given data. On each of times steps size observations are removed from the data, the functional DD-classifier is trained on these data and tested on the removed observations.

Usage

ddalphaf.getErrorRatePart(dataf, labels, size = 0.3, times = 10, 
                          disc.type = c("LS", "comp"),  ...)

Arguments

dataf
list containing lists (functions) of two vectors of equal length, named "args" and "vals": arguments sorted in ascending order and corresponding them values respectively

labels
list of output labels of the functional observations

size
the excluded sequences size. Either an integer between 1 and n, or a fraction of data between 0 and 1.

times
the number of times the classifier is trained.

disc.type
type of the used discretization scheme. "LS" for ddalphaf.train, "comp" for for compclassf.train

...
additional parameters passed to ddalphaf.train

Value

errors
the part of incorrectly classified data (mean)

errors_sd
the standard deviation of errors

errors_vec
vector of errors

time
the mean training time

time_sd
the standard deviation of training time

See Also

ddalphaf.train to train the functional DDα-classifier, ddalphaf.classify for classification using functional DDα-classifier, ddalphaf.test to test the functional DD-classifier on particular learning and testing data, ddalphaf.getErrorRateCV to get error rate of the functional DD-classifier on particular data.

Examples

# load the fdata
df = dataf.growth()

stat <- ddalphaf.getErrorRatePart(dataf = df$dataf, labels = df$labels, 
                          size = 0.3, times = 5,
                          adc.args = list(instance = "avr", 
                                         numFcn = 2, 
                                         numDer = 2))

cat("Classification error rate: ", stat$errors, ".\n", sep = "")

[Package ddalpha version 1.3.1.1 Index]