ddalphaf.test {ddalpha}	R Documentation

Test Functional DD-Classifier

Description

Trains functional DD-classifier on the learning sequence of the data and tests it on the testing sequence.

Usage

ddalphaf.test(learn, learnlabels, test, testlabels, disc.type = c("LS", "comp"), ...)

Arguments

learn
list containing lists (functions) of two vectors of equal length, named "args" and "vals": arguments sorted in ascending order and corresponding them values respectively

learnlabels
list of output labels of the functional observations

test
the testing sequence. Has the same format as learn

disc.type
type of the used discretization scheme. "LS" for ddalphaf.train, "comp" for for compclassf.train

testlabels
list of output labels of the functinal observations

...
additional parameters passed to ddalphaf.train

Value

error
the part of incorrectly classified data

correct
the number of correctly classified objects

incorrect
the number of incorrectly classified objects

total
the number of classified objects

ignored
the number of ignored objects (outside the convex hull of the learning data)

n
the number of objects in the testing sequence

time
training time

See Also

ddalphaf.train to train the functional DDα-classifier, ddalphaf.classify for classification using functonal DDα-classifier, ddalphaf.getErrorRateCV and ddalphaf.getErrorRatePart to get error rate of the functional DD-classifier on particular data.

Examples

# load the fdata
df = dataf.growth()

samp = c(35:70)

ddalphaf.test(learn = df$dataf[-samp], learnlabels = df$labels[-samp], 
              test =  df$dataf[samp],  testlabels =  df$labels[samp], 
              adc.args = list(instance = "avr", 
                              numFcn = 2, 
                              numDer = 2))

[Package ddalpha version 1.3.1.1 Index]