depth.L2 {ddalpha}	R Documentation

Calculate L2-Depth

Description

Calculates the L2-depth of points w.r.t. a multivariate data set.

Usage

depth.L2(x, data, mah.estimate = "moment", mah.parMcd = 0.75)

Arguments

x
Matrix of objects (numerical vector as one object) whose depth is to be calculated; each row contains a d-variate point. Should have the same dimension as data.

data
Matrix of data where each row contains a d-variate point, w.r.t. which the depth is to be calculated.

mah.estimate
is a character string specifying which estimates to use when calculating sample covariance matrix; can be "none", "moment" or "MCD", determining whether traditional moment or Minimum Covariance Determinant (MCD) (see covMcd) estimates for mean and covariance are used. By default "moment" is used. With "none" the non-affine invariant version of the L2-depth is calculated

mah.parMcd
is the value of the argument alpha for the function covMcd; is used when mah.estimate = "MCD".

Details

Calculates L2-depth (Mosler, 2013). L2-depth is a based on the oultyingness distance calculated as the average L2-distance from (a row of) x to each point in data.

Value

Numerical vector of depths, one for each row in x; or one depth value if x is a numerical vector.

References

Mosler, K. (2013). Depth statistics. In: Becker, C., Fried, R. and Kuhnt, S. (eds), Robustness and Complex Data Structures: Festschrift in Honour of Ursula Gather, Springer-Verlag (Berlin, Heidelberg), 17–34.

See Also

depth.halfspace for calculation of the Tukey depth.

depth.Mahalanobis for calculation of Mahalanobis depth.

depth.projection for calculation of projection depth.

depth.qhpeeling for calculation of convex hull peeling depth.

depth.simplicial for calculation of simplicial depth.

depth.simplicialVolume for calculation of simplicial volume depth.

depth.spatial for calculation of spatial depth.

depth.potential for calculation of data potential.

depth.zonoid for calculation of zonoid depth.

Examples

# 5-dimensional normal distribution
data <- mvrnorm(1000, rep(0, 5), 
                matrix(c(1, 0, 0, 0, 0, 
                         0, 2, 0, 0, 0, 
                         0, 0, 3, 0, 0, 
                         0, 0, 0, 2, 0, 
                         0, 0, 0, 0, 1),
                nrow = 5))
x <- mvrnorm(10, rep(1, 5), 
             matrix(c(1, 0, 0, 0, 0, 
                      0, 1, 0, 0, 0, 
                      0, 0, 1, 0, 0, 
                      0, 0, 0, 1, 0, 
                      0, 0, 0, 0, 1),
             nrow = 5))
                
depths <- depth.spatial(x, data)
cat("Depths:", depths, "\n")

[Package ddalpha version 1.3.8 Index]