depth.space.simplicialVolume {ddalpha}	R Documentation

Calculate Depth Space using Simplicial Volume Depth

Description

Calculates the representation of the training classes in depth space using simplicial volume depth.

Usage

depth.space.simplicialVolume(data, cardinalities, exact = F, k = 0.05, seed = 0)

Arguments

data
Matrix containing training sample where each row is a d-dimensional object, and objects of each class are kept together so that the matrix can be thought of as containing blocks of objects representing classes.

cardinalities
Numerical vector of cardinalities of each class in data, each entry corresponds to one class.

exact
exact=F (by default) implies the approximative algorithm, considering k simplices, exact=T implies the exact algorithm.

k
Number (k>1) or portion (if 0<k<1) of simplices that are considered if exact=F. If k>1, then the algorithmic complexity is polynomial in d but is independent of the number of observations in data, given k. If 0<k<1, then the algorithmic complexity is exponential in the number of observations in data, but the calculation precision stays approximately the same.

seed
The random seed. The default value seed=0 makes no changes.

Details

The depth representation is calculated in the same way as in depth.simplicialVolume, see References below for more information and details.

Value

Matrix of objects, each object (row) is represented via its depths (columns) w.r.t. each of the classes of the training sample; order of the classes in columns corresponds to the one in the argument cardinalities.

References

Oja, H. (1983). Descriptive statistics for multivariate distributions. Statistics & Probability Letters 1 327–332.

Zuo, Y.J. and Serfling, R. (2000). General notions of statistical depth function. The Annals of Statistics 28 461–482.

See Also

ddalpha.train and ddalpha.classify for application, depth.simplicialVolume for calculation of simplicial depth.

Examples

# Generate a bivariate normal location-shift classification task
# containing 20 training objects
class1 <- mvrnorm(10, c(0,0), 
                  matrix(c(1,1,1,4), nrow = 2, ncol = 2, byrow = TRUE))
class2 <- mvrnorm(10, c(2,2), 
                  matrix(c(1,1,1,4), nrow = 2, ncol = 2, byrow = TRUE))
data <- rbind(class1, class2)
# Get depth space using Oja depth
depth.space.simplicialVolume(data, c(10, 10))

data <- getdata("hemophilia")
cardinalities = c(sum(data$gr == "normal"), sum(data$gr == "carrier"))
depth.space.simplicialVolume(data[,1:2], cardinalities)

[Package ddalpha version 1.3.1.1 Index]