depthf. {ddalpha}	R Documentation

Calculate Functional Depth

Description

Calculates the depth of functions w.r.t. a functional data set.

The detailed descriptions are found in the corresponding topics.

Usage

depthf.(datafA, datafB, notion, ...)

## Adjusted band depth
# depthf.ABD(datafA, datafB, range = NULL, d = 101, norm = c("C", "L2"), 
# J = 2, K = 1)

## Band depth
# depthf.BD(datafA, datafB, range = NULL, d = 101)

## Univariate integrated and infimal depth
# depthf.fd1(datafA, datafB, range = NULL, d = 101, order = 1, approx = 0)

## Bivariate integrated and infimal depth
# depthf.fd2(datafA, datafB, range = NULL, d = 101)

## h-mode depth
# depthf.hM(datafA, datafB, range = NULL, d = 101, norm = c("C", "L2"),
#  q = 0.2)

## Bivariate h-mode depth
# depthf.hM2(datafA, datafB, range = NULL, d = 101, q = 0.2)

## Half-region depth
# depthf.HR(datafA, datafB, range = NULL, d = 101)

## Univariate random projection depths
# depthf.RP1(datafA, datafB, range = NULL, d = 101, nproj = 50, nproj2 = 5)

# Bivariate random projection depths
# depthf.RP2(datafA, datafB, range = NULL, d = 101, nproj = 51)

Arguments

datafA
Functions whose depth is computed, represented by a dataf object of their arguments and functional values.

datafB
Random sample functions with respect to which the depth of datafA is computed. datafB is represented by a dataf object of their arguments and functional values.

notion
The name of the depth notion (shall also work with a user-defined depth function named "depthf.<name>").

...
Additional parameters passed to the depth functions.

Value

Numerical vector of depths, one for each function in datafA; or one depth value if datafA is a single function.

See Also

depthf.ABD

depthf.BD

depthf.fd1

depthf.fd2

depthf.hM

depthf.hM2

depthf.HR

depthf.RP1

depthf.RP2

Examples

# real data example
datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]
                
depthf.(datafA, datafB, notion = "HR")

dataf2A = derivatives.est(datafA,deriv=c(0,1))
dataf2B = derivatives.est(datafB,deriv=c(0,1))

depthf.(dataf2A, dataf2B, notion = "fd2")

[Package ddalpha version 1.3.1.1 Index]