depthf.ABD {ddalpha}	R Documentation

Adjusted Band Depth for Functional Data

Description

The adjusted band depth of functional real-valued data based on either the C (uniform) norm, or on the L^2 norm of functions.

Usage

depthf.ABD(datafA, datafB, range = NULL, d = 101, norm = c("C", "L2"),
  J = 2, K = 1)

Arguments

datafA
Functions whose depth is computed, represented by a dataf object of their arguments and functional values. m stands for the number of functions.

datafB
Random sample functions with respect to which the depth of datafA is computed. datafB is represented by a dataf object of their arguments and functional values. n is the sample size. The grid of observation points for the functions datafA and datafB may not be the same.

range
The common range of the domain where the functions datafA and datafB are observed. Vector of length 2 with the left and the right end of the interval. Must contain all arguments given in datafA and datafB.

d
Grid size to which all the functional data are transformed. For depth computation, all functional observations are first transformed into vectors of their functional values of length d corresponding to equi-spaced points in the domain given by the interval range. Functional values in these points are reconstructed using linear interpolation, and extrapolation, see Nagy et al. (2016).

norm
The norm used for the computation of the depth. Two possible choices are implemented: C for the uniform norm of continuous functions, and L2 for the L^2 norm of integrable functions.

J
The order of the adjusted band depth, that is the maximal number of functions taken in a band. Acceptable values are 2, 3,... By default this value is set to 2. Note that this is NOT the order as defined in the order-extended version of adjusted band depths in Nagy et al. (2016), used for the detection of shape outlying curves.

K
Number of sub-samples of the functions from B taken to speed up the computation. By default, sub-sampling is not performed. Values of K larger than 1 result in an approximation of the adjusted band depth.

Details

The function returns the vector of the sample adjusted band depth values. The kernel used in the evaluation is the function K(u) = exp(-u).

Value

A vectors of length m of the adjusted band depths.

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

References

Gijbels, I., Nagy, S. (2015). Consistency of non-integrated depths for functional data. Journal of Multivariate Analysis 140, 259–282.

Nagy, S., Gijbels, I. and Hlubinka, D. (2016). Weak convergence of discretely observed functional data with applications. Journal of Multivariate Analysis, 146, 46–62.

Nagy, S., Gijbels, I. and Hlubinka, D. (2017). Depth-based recognition of shape outlying functions. Journal of Computational and Graphical Statistics. To appear.

See Also

depthf.BD

Examples

datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]
depthf.ABD(datafA,datafB)
depthf.ABD(datafA,datafB,norm="L2")

[Package ddalpha version 1.3.1.1 Index]