depthf.HR {ddalpha}	R Documentation

Half-Region Depth for Functional Data

Description

The half-region depth for functional real-valued data.

Usage

depthf.HR(datafA, datafB, range = NULL, d = 101)

Arguments

datafA
Functions whose depth is computed, represented by a dataf object of their arguments and functional values. m stands for the number of functions.

datafB
Random sample functions with respect to which the depth of datafA is computed. datafB is represented by a dataf object of their arguments and functional values. n is the sample size. The grid of observation points for the functions datafA and datafB may not be the same.

range
The common range of the domain where the functions datafA and datafB are observed. Vector of length 2 with the left and the right end of the interval. Must contain all arguments given in datafA and datafB.

d
Grid size to which all the functional data are transformed. For depth computation, all functional observations are first transformed into vectors of their functional values of length d corresponding to equi-spaced points in the domain given by the interval range. Functional values in these points are reconstructed using linear interpolation, and extrapolation.

Details

The function returns the vector of the sample half-region depth values.

Value

A vector of length m of the half-region depth values.

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

References

Lopez-Pintado, S. and Romo, J. (2011). A half-region depth for functional data. Computational Statistics & Data Analysis 55 (4), 1679–1695.

Examples

datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]
depthf.HR(datafA,datafB)

[Package ddalpha version 1.3.1.1 Index]