depthf.RP2 {ddalpha}	R Documentation

Bivariate Random Projection Depths for Functional Data

Description

Double random projection depths of functional bivariate data (that is, data of the form X:[a,b] \to R^2, or X:[a,b] \to R and the derivative of X).

Usage

depthf.RP2(datafA, datafB, range = NULL, d = 101, nproj = 51)

Arguments

datafA
Bivariate functions whose depth is computed, represented by a multivariate dataf object of their arguments (vector), and a matrix with two columns of the corresponding bivariate functional values. m stands for the number of functions.

datafB
Bivariate random sample functions with respect to which the depth of datafA is computed. datafB is represented by a multivariate dataf object of their arguments (vector), and a matrix with two columns of the corresponding bivariate functional values. n is the sample size. The grid of observation points for the functions datafA and datafB may not be the same.

range
The common range of the domain where the functions datafA and datafB are observed. Vector of length 2 with the left and the right end of the interval. Must contain all arguments given in datafA and datafB.

d
Grid size to which all the functional data are transformed. For depth computation, all functional observations are first transformed into vectors of their functional values of length d corresponding to equi-spaced points in the domain given by the interval range. Functional values in these points are reconstructed using linear interpolation, and extrapolation.

nproj
Number of projections taken in the computation of the double random projection depth. By default taken to be 51.

Details

The function returns the vectors of sample double random projection depth values. The double random projection depths are described in Cuevas et al. (2007). They are of two types: RP2 type, and RPD type. Both types of depths are based on bivariate projections of the bivariate functional data. These projections are taken randomly as a sample of standard normal d-dimensional random variables, where d stands for the dimensionality of the internally represented discretized functional data. For RP2 type depths, the average bivariate depth of the projected quantities is assessed. For RPD type depths, further univariate projections of these bivariate projected quantities are evaluated, and based on these final univariate quantities, the average univariate depth is computed.

Value

Five vectors of length m are returned:

Simpl_FD the double random projection depth RP2 based on the bivariate simplicial depth,

Half_FD the double random projection depth RP2 based on the bivariate halfspace depth,

hM_FD the double random projection depth RP2 based on the bivariate h-mode depth,

Simpl_DD the double random projection depth RPD based on the univariate simplicial depth,

Half_DD the random projection depth RPD based on the univariate halfspace depth,

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

References

Cuevas, A., Febrero, M. and Fraiman, R. (2007). Robust estimation and classification for functional data via projection-based depth notions. Computational Statistics 22 (3), 481–496.

See Also

depthf.RP1

Examples

datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]

dataf2A = derivatives.est(datafA,deriv=c(0,1))
dataf2B = derivatives.est(datafB,deriv=c(0,1))
depthf.RP2(dataf2A,dataf2B)

[Package ddalpha version 1.3.1.1 Index]