depthf.fd2 {ddalpha}	R Documentation

Bivariate Integrated and Infimal Depth for Functional Data

Description

Integrated and infimal depths of functional bivariate data (that is, data of the form X:[a,b] \to R^2, or X:[a,b] \to R and the derivative of X) based on the bivariate halfspace and simplicial depths.

Usage

depthf.fd2(datafA, datafB, range = NULL, d = 101)

Arguments

datafA
Bivariate functions whose depth is computed, represented by a multivariate dataf object of their arguments (vector), and a matrix with two columns of the corresponding bivariate functional values. m stands for the number of functions.

datafB
Bivariate random sample functions with respect to which the depth of datafA is computed. datafB is represented by a multivariate dataf object of their arguments (vector), and a matrix with two columns of the corresponding bivariate functional values. n is the sample size. The grid of observation points for the functions datafA and datafB may not be the same.

range
The common range of the domain where the functions datafA and datafB are observed. Vector of length 2 with the left and the right end of the interval. Must contain all arguments given in datafA and datafB.

d
Grid size to which all the functional data are transformed. For depth computation, all functional observations are first transformed into vectors of their functional values of length d corresponding to equi-spaced points in the domain given by the interval range. Functional values in these points are reconstructed using linear interpolation, and extrapolation.

Details

The function returns the vectors of sample integrated and infimal depth values.

Value

Four vectors of length m are returned:

Simpl_FD the integrated depth based on the bivariate simplicial depth,

Half_FD the integrated depth based on the bivariate halfspace depth,

Simpl_ID the infimal depth based on the bivariate simplicial depth,

Half_ID the infimal depth based on the bivariate halfspace depth.

In addition, two vectors of length m of the relative area of smallest depth values is returned:

Simpl_IA the proportions of points at which the depth Simpl_ID was attained,

Half_IA the proportions of points at which the depth Half_ID was attained.

The values Simpl_IA and Half_IA are always in the interval [0,1]. They introduce ranking also among functions having the same infimal depth value - if two functions have the same infimal depth, the one with larger infimal area IA is said to be less central.

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

References

Hlubinka, D., Gijbels, I., Omelka, M. and Nagy, S. (2015). Integrated data depth for smooth functions and its application in supervised classification. Computational Statistics, 30 (4), 1011–1031.

Nagy, S., Gijbels, I. and Hlubinka, D. (2017). Depth-based recognition of shape outlying functions. Journal of Computational and Graphical Statistics. To appear.

See Also

depthf.fd1, infimalRank

Examples

datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]

dataf2A = derivatives.est(datafA,deriv=c(0,1))
dataf2B = derivatives.est(datafB,deriv=c(0,1))
depthf.fd2(dataf2A,dataf2B)

[Package ddalpha version 1.3.1.1 Index]