depthf.hM2 {ddalpha}	R Documentation

Bivariate h-Mode Depth for Functional Data Based on the L^2 Metric

Description

The h-mode depth of functional bivariate data (that is, data of the form X:[a,b] \to R^2, or X:[a,b] \to R and the derivative of X) based on the L^2 metric of functions.

Usage

depthf.hM2(datafA, datafB, range = NULL, d = 101, q = 0.2)

Arguments

datafA
Bivariate functions whose depth is computed, represented by a multivariate dataf object of their arguments (vector), and a matrix with two columns of the corresponding bivariate functional values. m stands for the number of functions.

datafB
Bivariate random sample functions with respect to which the depth of datafA is computed. datafB is represented by a multivariate dataf object of their arguments (vector), and a matrix with two columns of the corresponding bivariate functional values. n is the sample size. The grid of observation points for the functions datafA and datafB may not be the same.

range
The common range of the domain where the functions datafA and datafB are observed. Vector of length 2 with the left and the right end of the interval. Must contain all arguments given in datafA and datafB.

d
Grid size to which all the functional data are transformed. For depth computation, all functional observations are first transformed into vectors of their functional values of length d corresponding to equi-spaced points in the domain given by the interval range. Functional values in these points are reconstructed using linear interpolation, and extrapolation.

q
The quantile used to determine the value of the bandwidth h in the computation of the h-mode depth. h is taken as the q-quantile of all non-zero distances between the functions B. By default, this value is set to q=0.2, in accordance with the choice of Cuevas et al. (2007).

Details

The function returns the vectors of sample h-mode depth values. The kernel used in the evaluation is the standard Gaussian kernel, the bandwidth value is chosen as a quantile of the non-zero distances between the random sample curves.

Value

Three vectors of length m of h-mode depth values are returned:

hM the unscaled h-mode depth,

hM_norm the h-mode depth hM linearly transformed so that its range is [0,1],

hM_norm2 the h-mode depth FD linearly transformed by a transformation such that the range of the h-mode depth of B with respect to B is [0,1]. This depth may give negative values.

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

References

Cuevas, A., Febrero, M. and Fraiman, R. (2007). Robust estimation and classification for functional data via projection-based depth notions. Computational Statistics 22 (3), 481–496.

See Also

depthf.hM

Examples

datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]

datafA2 = derivatives.est(datafA,deriv=c(0,1))
datafB2 = derivatives.est(datafB,deriv=c(0,1))

depthf.hM2(datafA2,datafB2)

depthf.hM2(datafA2,datafB2)$hM
# depthf.hM2(cbind(A2[,,1],A2[,,2]),cbind(B2[,,1],B2[,,2]))$hM
# the two expressions above should give the same result

[Package ddalpha version 1.3.1.1 Index]