depthf.simplicialBand {ddalpha}	R Documentation

Calculate Simplicial Band Depth

Description

Calculate the simplicial band depth defined by Lopez-Pintado, Sun, Lin, Genton (2014).

Usage

depthf.simplicialBand(objectsf, dataf, modified = TRUE, J = NULL, 
                      range = NULL, d = 101)

Arguments

objectsf
Functoins for which the depth should be computed; a list containing lists (functions) of two vectors of equal length, named args and vals: arguments sorted in ascending order and corresponding them values respectively.

dataf
Data sample of functoins w.r.t. which the depth should be computed; structure as for objectsf.

modified
Whether modified simplicial band depth should be computed; logical, TRUE by default.

J
How many functions to consider in each tuple of the U-statistics; integer, d+1 by default.

range
The common range of the domain where the functions of objectsf and dataf are observed. Vector of length 2 with the left and the right end of the interval. Must contain all arguments given in objectsf and dataf.

d
Grid size to which all the functional data are transformed. For depth computation, all functional observations are first transformed into vectors of their functional values of length d corresponding to equi-spaced points in the domain given by the interval range. Functional values in these points are reconstructed using linear interpolation, and extrapolation.

Value

A vector of depths of each of objectsf w.r.t. dataf.

References

Lopez-Pintado, Sun, Lin, Genton (2014). Simplicial band depth for multivariate data. Advances in Data Analysis and Classification 8(3) 321–338.

See Also

depthf.BD, depthf.ABD.

Examples

datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]

dataf2A = derivatives.est(datafA,deriv=c(0,1))
dataf2B = derivatives.est(datafB,deriv=c(0,1))
depthf.simplicialBand(dataf2A,dataf2B)

[Package ddalpha version 1.3.10 Index]