derivatives.est {ddalpha}	R Documentation

Estimation of the First Two Derivatives for Functional Data

Description

Returns the estimated values of derivatives of functional data.

Usage

derivatives.est(dataf, range = NULL, d = 101, spar = NULL, deriv = c(0,
  1))

Arguments

dataf
Functional dataset, represented by a dataf object of their arguments and functional values. m stands for the number of functions.

range
The common range of the domain where the functions dataf are observed. Vector of length 2 with the left and the right end of the interval. Must contain all arguments given in dataf.

d
Grid size to which all the functional data are transformed. For computation, all functional observations are first transformed into vectors of their functional values of length d corresponding to equi-spaced points in the domain given by the interval range. Functional values in these points are reconstructed using linear interpolation, and extrapolation.

spar
If provided, this parameter is passed to functions D1ss and D2ss from package sfsmisc as the value of the smoothing spline parameter in order to numerically approximate the derivatives of dataf.

deriv
A vector composed of 0, 1, and 2 of the demanded functional values / derivatives of the functions in the rows of dataf. 0 stands for the functional values, 1 for the first derivatives, 2 for the second derivatives.

Details

If the input dataf is a functional random sample of size m, the function returns a dataf object of nd-dimensional functional data, where in the elements of the vector-valued functional data represent the estimated values of the derivatives of dataf. All derivatives are evaluated at an equi-distant grid of d points in the domain given by range. nd here stands for 1, 2 or 3, depending on how many derivatives of dataf are requested to be computed. For the estimation, functions D1ss and D2ss from the package sfsmisc are utilized.

Value

A multivariate dataf object of the functional values and / or the derivatives of dataf. The dimensionality of the vector-valued functional data is nd. The arguments of the data are all equal to an equi-distant grid of d points in the domain given by range. nd is the demanded number of derivatives at the output, i.e. the length of the vector deriv.

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

See Also

D1ss in package sfsmisc

D2ss in package sfsmisc

Examples

dataf = dataf.population()$dataf
derivatives.est(dataf,deriv=c(0,1,2))

[Package ddalpha version 1.3.1.1 Index]