infimalRank {ddalpha}	R Documentation

Adjusted Ranking of Functional Data Based on the Infimal Depth

Description

Returns a vector of adjusted depth-based ranks for infimal depth for functional data.

Usage

infimalRank(ID, IA, ties.method = "max")

Arguments

ID
The vector of infimal depths of the curves of length n.

IA
The vector of the infimal areas corresponding to the infimal depths from ID of length n.

ties.method
Parameter for breaking ties in infimal area index. By default max, see rank.

Details

Infimal depths for functional data tend to give to many functional observations the same value of depth. Using this function, the data whose depth is the same is ranked according to the infimal area indicator. This indicator is provided in functions depthf.fd1 along the value of the infimal depth.

Value

A vector of length n. Low depth values mean high ranks, i.e. potential outlyingness. If some of the infimal depths are identical, the ranking of these functions is made according to the values of the infimal area. There, higher infimal area index means higher rank, i.e. non-centrality.

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

References

Nagy, S., Gijbels, I. and Hlubinka, D. (2017). Depth-based recognition of shape outlying functions. Journal of Computational and Graphical Statistics. To appear.

Examples

datafA = dataf.population()$dataf[1:20]
datafB = dataf.population()$dataf[21:50]
D = depthf.fd1(datafA,datafB)
infimalRank(D$Half_ID,D$Half_IA) 

ID = c(0,1,0,0,0,1,1)
IA = c(2,3,1,0,2,4,1)
infimalRank(ID,IA)

[Package ddalpha version 1.3.1.1 Index]