plot.ddalphaf {ddalpha}	R Documentation

Plots for the "ddalphaf" Class

Description

plot.functional – plots the functional data used by classifier

depth.contours.ddalpha – builds the data depth contours for multiclass 2-dimensional data using the trained classifier. draw.ddplot – draws the DD-plot of the existing DDα-classifier.

Usage

## S3 method for class 'ddalphaf'
plot(x, type = c("functional.data", "ddplot", "depth.contours"), ...)

Arguments

x
functional DDα-classifier (obtained by ddalphaf.train).

type
type of the plot for plot.functional, draw.ddplot or depth.contours.ddalpha

...
additional parameters passed to the depth functions and to plot

See Also

depth.

depth.contours

depth.graph

Examples

## Not run: 

dataf = dataf.growth()
ddalphaf = ddalphaf.train (dataf$dataf, dataf$labels, 
                            classifier.type = "ddalpha", maxNumIntervals = 2)

# plot the functional data
plot(ddalphaf)

# plot depth contours and separation in the transformed space 
# (possible only if maxNumIntervals = 2)
plot(ddalphaf, type = "depth.contours")

# plot the DD-plot
plot(ddalphaf, type = "ddplot")

## End(Not run)

[Package ddalpha version 1.3.1.1 Index]