rawfd2dataf {ddalpha}	R Documentation

Transform Raw Functional Data to a dataf Object

Description

Constructs a (possibly multivariate) functional data object given by an array of its functional values evaluated at an equi-distant grid of points, and transforms it into a dataf object more suitable for work in the ddalpha package.

Usage

rawfd2dataf(X, range)

Arguments

X
Either a matrix of size n*d, or an array of dimension n*d*k of functional values. Here n stands for the number of functions, d is the number of equi-distant points in the domain where the functional values are evaluated, and if applicable, k is the dimensionality of the (vector-valued) functional data.

range
A vector of size two that represents the endpoints of the common domain of all functions X.

Value

A (possibly multivariate) dataf object corresponding to the functional data X evaluated at an equi-distant grid of points.

Author(s)

Stanislav Nagy, nagy at karlin.mff.cuni.cz

See Also

dataf2rawfd

depthf.fd1

depthf.fd2

Examples

## transform a matrix into a functional data set
n = 5
d = 21
X = matrix(rnorm(n*d),ncol=d)
rawfd2dataf(X,range=c(0,1))

## transform an array into a multivariate functional data set
k = 3
X = array(rnorm(n*d*k),dim=c(n,d,k))
rawfd2dataf(X,range=c(-1,1))

[Package ddalpha version 1.3.1.1 Index]