resetPar {ddalpha}	R Documentation

Reset Graphical Parameters

Description

The function returns the default graphical parameters for par.

Usage

resetPar()

Details

The returned parameters are used as input parameters for par.

Value

The list of graphical parameters described in the 'Graphical Parameters' section of par.

Examples

par(mfrow = c(1,2), mar = c(2,2,2,2))
plot(sin, -pi, 2*pi)
plot(cos, -pi, 2*pi)

par(resetPar())
plot(sin, -pi, 2*pi)
plot(cos, -pi, 2*pi)

[Package ddalpha version 1.3.1.1 Index]